Add files via upload
A drunk man wants to go home. Now, he is in front of the bar and his home is on the opposite side of a circular street. Suppose he is standing at the 50th meter of the street and you can denote the location of his home as at the 100th or the 0th meter of the road.
He is so drunk that he steps forward with a possibility of 50% and another 50% backward. The step size of the man is 1 meter.
Write a Monte-Carlo random walk simulation and find how many steps should the drunken man walk to get his home. Plot his position versus steps ( i th position - i th step).